+++ utilities/define/def_schism_transect.m
@@ -18,20 +18,20 @@ function sect_info = def_schism_transect(Mobj, mtype, dx)
 %       the mesh object with vertical grid info added.
 % mtype - method type; numeric
 %       the flag used to specify the defning method of transect. 
-%       Four different options are available. Default: mtype = 1;   
-%       -1: draw a line on the map and return the vertical layers strictly
-%            along the transect (transect points are not on the node center).
+%       Five different options are available. Default: mtype = 1;   
 %        1: draw a line on the map and return the vertical lalyers on the
 %            nearset nodes to the transect points.
 %        2: click some node centers on the map using datatip and return the
 %            vertical layers on these nodes.
 %        3: draw broken lines on the map using polyline and return the
 %            vertical layers strickly along the transect (transect points
-%            are not on the node center). 
+%            are on the node center).
+%       -1: the same as option "1" but the points are not on node centers.
+%       -3: the same as option "3" but the points are not on node centers.
 % dx - the density of transect points; numeric
 %       if mtype =-1, 1, or 3: dx can be used to determine the density of
 %       transect points, ranging from 0 to 1, and lower values mean high
-%       density. Default: dx = 0.02; 
+%       density. Default: dx = 0.02;
 %       if mtype = 2: dx =1 means the transect points will be arranged in
 %       a descending order based on the latitudes, while -1 means an
 %       ascending order. Default: dx = 1;
@@ -66,26 +66,27 @@ if nargin<3
 end
 %% Define transect
 switch mtype
-    case -1 % drawline (will not find nearest nodes)
+    case 1 %  straight line (points are on the node center)
         disp('Please draw a line on the map and press ENTER to end')
-        sect_handle = drawline; pause;
-        lon_roi = sect_handle.Position(:,1)'; lat_roi = sect_handle.Position(:,2)'; close gcf
-        lon_list = interp1(1:2, lon_roi, 1:dx:2); lat_list = interp1(1:2, lat_roi, 1:dx:2);
-        F = scatteredInterpolant(Mobj.lon, Mobj.lat, Mobj.depth);
-        dep_list = F(lon_list, lat_list);
-        ind_nodes = minfind(Mobj.depth, dep_list);  % use depth layers from the nearest depth.
-
-    case 1 % drawline (find the nearest nodes on the mesh)
-        disp('Please draw a line on the map and press ENTER to end')
-        sect_handle = drawline; 
+        sect_handle = drawline;
         pause;
         lon_roi = sect_handle.Position(:,1)'; lat_roi = sect_handle.Position(:,2)';
         close gcf
-        lon_roi_ref = interp1(1:2, lon_roi, 1:dx:2); lat_roi_ref = interp1(1:2, lat_roi, 1:dx:2);  % refine the transect points
-        ind_nodes = unique(geomin(Mobj.lon, Mobj.lat, lon_roi_ref, lat_roi_ref), 'stable');
-        lon_list = Mobj.lon(ind_nodes); lat_list = Mobj.lat(ind_nodes);  % find the nearest nodes
+        lon_roi2 = interp1(1:2, lon_roi, 1:dx:2); lat_roi2 = interp1(1:2, lat_roi, 1:dx:2);  % refine the transect points
+        ind_nodes = unique(geomin(Mobj.lon, Mobj.lat, lon_roi2, lat_roi2), 'stable'); % find the nearest nodes
+        lon_list = Mobj.lon(ind_nodes); lat_list = Mobj.lat(ind_nodes);
 
-    case 2  % datatip
+    case -1 % straight line (points are not on the node center)
+        disp('Please draw a line on the map and press ENTER to end')
+        sect_handle = drawline;
+        pause;
+        lon_roi = sect_handle.Position(:,1)'; lat_roi = sect_handle.Position(:,2)';
+        close gcf
+        lon_list = interp1(1:2, lon_roi, 1:dx:2); lat_list = interp1(1:2, lat_roi, 1:dx:2);
+        F = scatteredInterpolant(Mobj.lon, Mobj.lat, Mobj.depth);
+        ind_nodes = minfind(Mobj.depth, F(lon_list, lat_list));  % use depth layers from the nearest depth.
+
+    case 2  % scattered points (points are on the node center)
         disp('Please click all transect nodes on the map using datatips and press ENTER to end')
         pause
         dataTips = findobj(gcf,'Type','datatip');
@@ -93,44 +94,61 @@ switch mtype
             xy_data = cell2array(arrayfun(@(x) [dataTips(x).X; dataTips(x).Y], 1:length(dataTips), 'UniformOutput', false));
             close gcf
             lon_list = xy_data(:,1); lat_list =  xy_data(:,2);
+            nps = numel(lon_list);
         end
-        if dx == 1  % latitude-descending order 
+        if sign(dx) == 1  % latitude-descending order
             [lat_list, ind] = sort(lat_list, 'descend');
             lon_list = lon_list(ind);
-        elseif dx==-1 % latitude-ascending order 
+        elseif sign(dx) < 1 % latitude-ascending order
             [lat_list, ind] = sort(lat_list, 'ascend');
             lon_list = lon_list(ind);
         end
         ind_nodes = arrayfun(@(x) geomin(Mobj.lon, Mobj.lat, lon_list(x), lat_list(x)), 1:nps);
 
-    case 3  % polyline (may not work)
+    case 3  % broken line (points are on the node center)
+        disp('Please draw broken lines on the map and press ENTER to end')
         geo_handle = drawpolyline;
         lon_roi = geo_handle.Position(:,1)'; lat_roi = geo_handle.Position(:,2)';
         close gcf; clear geo_handle
-        lon_list = []; lat_list = [];
+        lon_roi2 = []; lat_roi2 = [];
         for s = 1:numel(lon_roi)-1
             lon_seg = interp1(1:2, lon_roi(s:s+1), 1:dx:2);
             lat_seg = interp1(1:2, lat_roi(s:s+1), 1:dx:2);
-            lon_list = [lon_list; lon_seg(:)]; lat_list = [lat_list; lat_seg(:)]; %#ok<AGROW>
+            lon_roi2 = [lon_roi2; lon_seg(:)]; lat_roi2 = [lat_roi2; lat_seg(:)]; %#ok<AGROW>
         end
-        [~, ia, ~] = unique(hypot(lon_list, lat_list), 'stable');
-        lon_list = lon_list(ia); lat_list = lat_list(ia);
+        [~, ia, ~] = unique(hypot(lon_roi2, lat_roi2), 'stable');  % remove duplicate points
+        lon_roi2 = lon_roi2(ia); lat_roi2 = lat_roi2(ia);
+        ind_nodes = unique(geomin(Mobj.lon, Mobj.lat, lon_roi2, lat_roi2), 'stable');
+        lon_list = Mobj.lon(ind_nodes); lat_list = Mobj.lat(ind_nodes);  % find the nearest nodes
+
+    case -3  % broken line (points are not on the node center)
+        disp('Please draw broken lines on the map and press ENTER to end')
+        geo_handle = drawpolyline;
+        lon_roi = geo_handle.Position(:,1)'; lat_roi = geo_handle.Position(:,2)';
+        close gcf; clear geo_handle
+        lon_roi2 = []; lat_roi2 = [];
+        for s = 1:numel(lon_roi)-1
+            lon_seg = interp1(1:2, lon_roi(s:s+1), 1:dx:2);
+            lat_seg = interp1(1:2, lat_roi(s:s+1), 1:dx:2);
+            lon_roi2 = [lon_roi2; lon_seg(:)]; lat_roi2 = [lat_roi2; lat_seg(:)]; %#ok<AGROW>
+        end
+        [~, ia, ~] = unique(hypot(lon_roi2, lat_roi2), 'stable');
+        lon_list = lon_roi2(ia); lat_list = lat_roi2(ia);
 
         F = scatteredInterpolant(Mobj.lon, Mobj.lat, Mobj.depth);
-        dep_list = F(lon_list, lat_list);
-        ind_nodes = minfind(Mobj.depth, dep_list);
+        ind_nodes = minfind(Mobj.depth, F(lon_list, lat_list));
 end
 %% Distance
 switch lower(Mobj.coord)
     case 'geographic'
-        dist_roi = distance(lat_list(1), lon_list(1), lat_list, lon_list, [6378137 0.0818191910428158]); % meters
+        dist_list = distance(lat_list(1), lon_list(1), lat_list, lon_list, [6378137 0.0818191910428158]); % meters
     case 'cartesian'
-        dist_roi = hypot(lon_list-lon_list(1), lat_list-lat_list(1));
+        dist_list = hypot(lon_list-lon_list(1), lat_list-lat_list(1));
 end
 %% Save transect info
-sect_info.lon = lon_list(:);
-sect_info.lat = lat_list(:);
-sect_info.dist = dist_roi(:);
+sect_info.lon = lon_list(:); 
+sect_info.lat = lat_list(:); 
+sect_info.dist = dist_list(:);
 
 % Find the depth layers.
 if isfield(Mobj, 'depLayers')
@@ -141,7 +159,7 @@ end
 if mtype == -1 || mtype == 1  % only for straight line.
     A = [diff(lon_roi) diff(lat_roi)]; 
     B = [1, 0];
-    sect_info.theta = vector_angle(A, B);
-    sect_info.vector = A;
+    sect_info.theta = vector_angle(A, B);   % anti-clockwise angle from A to B
+    sect_info.vector = A;    % transect vector
 end
 end